TheDescent.cpp
*Problem Statement:
Destroy the mountains before your starship collides with one of them. For that, shoot the highest mountain on your path.

*Rules
At the start of each game turn, you are given the height of the 8 mountains from left to right.
By the end of the game turn, you must fire on the highest mountain by outputting its index (from 0 to 7).
Firing on a mountain will only destroy part of it, reducing its height. Your ship descends after each pass.  
	
*Game Input
Within an infinite loop, read the heights of the mountains from the standard input and print to the standard output the index of the mountain to shoot.

*Output for one game turn
A single line with one integer for the index of which mountain to shoot.